+++ gui/util/string.h
@@ -27,7 +27,7 @@ int gstrcasecmp(const char *a, const char *b);
 char *gstrchr(const char *str, int c);
 int gstrcmp(const char *a, const char *b);
 char *gstrdup(const char *str);
-int gstrncmp(const char *a, const char *b, int n);
+int gstrncmp(const char *a, const char *b, size_t n);
 void setddup(char **old, const char *dir, const char *name);
 void setdup(char **old, const char *str);
 char *strlower(char *in);
